Modularized the code (part 1)
* Part 1 of modularizing the code
